move episode list files
